refactor: decompose repl.rs into submodules (H1)
Move `src/repl.rs` to `src/repl/mod.rs` and extract the AI
command handlers (~1 700 lines) into `src/repl/ai.rs`.

- `repl/mod.rs` — types, settings, prompt, query execution,
  backslash dispatch, readline loops, and tests (unchanged API)
- `repl/ai.rs` — all `/ask`, `/fix`, `/explain`, `/optimize`,
  `/describe`, `/rca`, and supporting helpers; accessed from
  `mod.rs` via explicit named imports

All 1 361 tests pass; `cargo clippy -D warnings` is clean.
Public API unchanged: callers still use `mod repl;` as before.

Closes #339 (H1)
