+++ src/repl/mod.rs
@@ -6595,1699 +6595,26 @@ async fn handle_line(
 }
 
 // ---------------------------------------------------------------------------
-// AI command helpers
+
+// ---------------------------------------------------------------------------
+// Submodule: AI command handlers
 // ---------------------------------------------------------------------------
 
-/// Stream an LLM completion to stdout, printing tokens as they arrive.
-///
-/// Falls back to printing the full response at once if the provider does
-/// not implement true streaming.
-/// Show a brief inline AI suggestion after a SQL error.
-///
-/// Called automatically when `[ai] auto_explain_errors = true`.  The
-/// suggestion is dimmed to visually distinguish it from the error itself.
-/// Uses a small `max_tokens` budget to keep latency low.
-async fn suggest_error_fix_inline(sql: &str, error_message: &str, settings: &mut ReplSettings) {
-    if check_token_budget(settings) {
-        return;
-    }
+pub(crate) mod ai;
 
-    let Ok(provider) = get_ai_provider(settings) else {
-        return; // AI not configured or error — silently skip.
-    };
+// Items from `ai` used in this module's production code.
+use ai::{
+    ask_yn_prompt, dispatch_ai_command, handle_ai_ask, handle_ai_plan, interpret_auto_explain,
+    interpret_dba_output, record_token_usage, stream_completion, suggest_error_fix_inline,
+};
 
-    let messages = vec![
-        crate::ai::Message {
-            role: crate::ai::Role::System,
-            content: "You are a PostgreSQL expert. \
-                      The user just got a SQL error. \
-                      Give a ONE-LINE fix suggestion. \
-                      Be extremely concise — just the fix, nothing else."
-                .to_owned(),
-        },
-        crate::ai::Message {
-            role: crate::ai::Role::User,
-            content: format!("Query: {sql}\nError: {error_message}"),
-        },
-    ];
-
-    let options = crate::ai::CompletionOptions {
-        model: settings.config.ai.model.clone().unwrap_or_default(),
-        max_tokens: 150,
-        temperature: 0.0,
-    };
-
-    // Use non-streaming for lower latency on a short response.
-    if let Ok(result) = provider.complete(&messages, &options).await {
-        record_token_usage(settings, &result);
-        let suggestion = result.content.trim();
-        if !suggestion.is_empty() {
-            // Print dimmed (ANSI escape: dim = \x1b[2m, reset = \x1b[0m).
-            eprintln!("\x1b[2mHint: {suggestion}\x1b[0m");
-        }
-    }
-}
-
-/// Interpret an auto-EXPLAIN plan with AI.
-///
-/// Called automatically after auto-EXPLAIN output is displayed. Uses a
-/// concise system prompt to produce a brief interpretation of the plan.
-/// Skips silently when AI is not configured or the token budget is exhausted.
-async fn interpret_auto_explain(
-    plan_text: &str,
-    original_query: &str,
-    settings: &mut ReplSettings,
-) {
-    if check_token_budget(settings) {
-        return;
-    }
-
-    let Ok(provider) = get_ai_provider(settings) else {
-        return; // AI not configured or error — silently skip.
-    };
-
-    let messages = vec![
-        crate::ai::Message {
-            role: crate::ai::Role::System,
-            content: "You are a PostgreSQL performance expert. \
-                      Give a brief (2-4 sentence) interpretation of the \
-                      EXPLAIN plan. Focus on: most expensive nodes, \
-                      sequential scans, row estimate errors, and one \
-                      actionable suggestion if applicable."
-                .to_owned(),
-        },
-        crate::ai::Message {
-            role: crate::ai::Role::User,
-            content: format!("Query: {original_query}\n\nPlan:\n{plan_text}"),
-        },
-    ];
-
-    let options = crate::ai::CompletionOptions {
-        model: settings.config.ai.model.clone().unwrap_or_default(),
-        max_tokens: 300,
-        temperature: 0.0,
-    };
-
-    if let Ok(result) = provider.complete(&messages, &options).await {
-        record_token_usage(settings, &result);
-        let interpretation = result.content.trim();
-        if !interpretation.is_empty() {
-            eprintln!("\x1b[2m{interpretation}\x1b[0m");
-        }
-    }
-}
-
-/// Interpret `\dba` diagnostic output with AI.
-///
-/// Called when a `\dba` command returns AI context (e.g. `\dba waits+`).
-/// Produces a brief analysis of the diagnostic data. Skips silently when
-/// AI is not configured.
-async fn interpret_dba_output(context: &str, subcommand: &str, settings: &mut ReplSettings) {
-    if check_token_budget(settings) {
-        return;
-    }
-
-    let Ok(provider) = get_ai_provider(settings) else {
-        eprintln!("-- AI interpretation requires [ai] provider to be configured");
-        return;
-    };
-
-    let messages = vec![
-        crate::ai::Message {
-            role: crate::ai::Role::System,
-            content: "You are a PostgreSQL performance expert. \
-                      Analyze the diagnostic output below and give a brief (3-5 sentence) \
-                      interpretation. Focus on: dominant wait events, potential bottlenecks, \
-                      and one actionable recommendation if applicable."
-                .to_owned(),
-        },
-        crate::ai::Message {
-            role: crate::ai::Role::User,
-            content: format!("\\dba {subcommand} output:\n\n{context}"),
-        },
-    ];
-
-    let options = crate::ai::CompletionOptions {
-        model: settings.config.ai.model.clone().unwrap_or_default(),
-        max_tokens: 400,
-        temperature: 0.0,
-    };
-
-    eprintln!("-- AI interpreting wait events...");
-    if let Ok(result) = provider.complete(&messages, &options).await {
-        record_token_usage(settings, &result);
-        let interpretation = result.content.trim();
-        if !interpretation.is_empty() {
-            eprintln!("\x1b[2m{interpretation}\x1b[0m");
-        }
-    }
-}
-
-/// Stream a completion to the terminal, rendering markdown when enabled.
-///
-/// When `no_highlight` is `false` (the default), tokens are buffered and
-/// the completed response is passed through [`crate::markdown::render_markdown`]
-/// before printing, so that headers, bold text, code fences, etc. are
-/// displayed with ANSI styling.
-///
-/// When `no_highlight` is `true` the raw text is streamed directly to
-/// stdout token by token (original behaviour).
-async fn stream_completion(
-    provider: &dyn crate::ai::LlmProvider,
-    messages: &[crate::ai::Message],
-    options: &crate::ai::CompletionOptions,
-    no_highlight: bool,
-) -> Result<crate::ai::CompletionResult, String> {
-    use std::io::Write;
-
-    if no_highlight {
-        // Raw streaming — emit each token immediately as it arrives.
-        let result = provider
-            .complete_streaming(
-                messages,
-                options,
-                Box::new(|token| {
-                    print!("{token}");
-                    let _ = io::stdout().flush();
-                }),
-            )
-            .await?;
-        println!();
-        return Ok(result);
-    }
-
-    // Markdown rendering mode — buffer tokens, render after completion.
-    //
-    // We still want to show progress to the user, so we print a dim "…"
-    // indicator that gets overwritten once the full response arrives.
-    // Use a shared buffer via Arc<Mutex<String>>.
-    let buf = std::sync::Arc::new(std::sync::Mutex::new(String::new()));
-    let buf_clone = buf.clone();
-
-    // Show a progress indicator so the terminal doesn't look stuck.
-    eprint!("\x1b[2m…\x1b[0m");
-    let _ = io::stderr().flush();
-
-    let result = provider
-        .complete_streaming(
-            messages,
-            options,
-            Box::new(move |token| {
-                if let Ok(mut b) = buf_clone.lock() {
-                    b.push_str(token);
-                }
-            }),
-        )
-        .await?;
-
-    // Erase the progress indicator (carriage return clears the line).
-    eprint!("\r\x1b[K");
-    let _ = io::stderr().flush();
-
-    // Render markdown on the fully-collected content and print.
-    let content = buf.lock().map(|b| b.clone()).unwrap_or_default();
-    let rendered = crate::markdown::render_markdown(&content, false);
-    print!("{rendered}");
-    let _ = io::stdout().flush();
-
-    Ok(result)
-}
-
-/// Dispatch a `/`-prefixed AI command.
-///
-/// Recognised commands:
-/// - `/ask <prompt>` — generate SQL from natural language
-/// - `/fix` — explain and fix the last error
-/// - `/explain [query]` — explain query plan with AI interpretation
-/// - `/optimize [query]` — suggest query optimizations
-/// - `/rca` — root cause analysis of current database state
-async fn dispatch_ai_command(
-    input: &str,
-    client: &Client,
-    params: &ConnParams,
-    settings: &mut ReplSettings,
-    tx: &mut TxState,
-) {
-    // Budget gate — skip for /clear and /compact (they don't use tokens).
-    let is_budget_exempt =
-        input == "/clear" || input.starts_with("/compact") || input.starts_with("/budget");
-    if !is_budget_exempt && check_token_budget(settings) {
-        return;
-    }
-
-    if let Some(prompt) = input.strip_prefix("/ask").map(str::trim) {
-        if prompt.is_empty() {
-            eprintln!("Usage: /ask <natural language description>");
-            return;
-        }
-        match settings.exec_mode {
-            ExecMode::Plan => handle_ai_plan(client, prompt, settings, params).await,
-            _ => handle_ai_ask(client, prompt, settings, params, tx).await,
-        }
-    } else if input == "/fix" || input.starts_with("/fix ") {
-        handle_ai_fix(client, settings, params, tx).await;
-    } else if let Some(query_arg) = input.strip_prefix("/explain").map(str::trim) {
-        handle_ai_explain(client, query_arg, settings, params).await;
-    } else if let Some(query_arg) = input.strip_prefix("/optimize").map(str::trim) {
-        handle_ai_optimize(client, query_arg, settings, params).await;
-    } else if let Some(table_arg) = input.strip_prefix("/describe").map(str::trim) {
-        if table_arg.is_empty() {
-            eprintln!("Usage: /describe <table_name>");
-            return;
-        }
-        handle_ai_describe(client, table_arg, settings, params).await;
-    } else if input == "/clear" {
-        settings.conversation.clear();
-        eprintln!("AI conversation context cleared.");
-    } else if let Some(focus) = input.strip_prefix("/compact").map(str::trim) {
-        if settings.conversation.is_empty() {
-            eprintln!("Nothing to compact — conversation context is empty.");
-        } else {
-            let focus = if focus.is_empty() { None } else { Some(focus) };
-            let before = settings.conversation.entries.len();
-            settings.conversation.compact(focus);
-            eprintln!(
-                "Compacted {before} entries → {} entries (~{} tokens)",
-                settings.conversation.entries.len(),
-                settings.conversation.token_estimate(),
-            );
-        }
-    } else if input == "/budget" {
-        let budget = settings.config.ai.token_budget;
-        let used = settings.tokens_used;
-        if budget == 0 {
-            eprintln!("Token budget: unlimited ({used} tokens used this session)");
-        } else {
-            let remaining = budget.saturating_sub(used);
-            eprintln!("Token budget: {used}/{budget} used, {remaining} remaining");
-        }
-    } else if input == "/rca" || input.starts_with("/rca ") {
-        handle_ai_rca(client, settings, params).await;
-    } else {
-        eprintln!(
-            "Unknown AI command: {input}\n\
-             Available: /ask, /fix, /explain, /optimize, /describe, /rca, /clear, /compact, /budget"
-        );
-    }
-}
-
-/// Strip markdown code fences from LLM output.
-///
-/// LLMs sometimes wrap SQL in `` ```sql ... ``` `` blocks.  This function
-/// removes the fences and returns the inner content, trimmed.  If no fences
-/// are found, the original string is returned as-is.
+// Items from `ai` used only in the `#[cfg(test)]` module below.
 #[cfg(test)]
-fn strip_sql_fences(s: &str) -> &str {
-    let trimmed = s.trim();
-    if let Some(rest) = trimmed.strip_prefix("```") {
-        // Skip optional language tag on the opening fence line.
-        let after_tag = rest.find('\n').map_or(rest, |i| &rest[i + 1..]);
-        // Remove closing fence.
-        let body = if let Some(pos) = after_tag.rfind("```") {
-            &after_tag[..pos]
-        } else {
-            after_tag
-        };
-        body.trim()
-    } else {
-        trimmed
-    }
-}
+use ai::{
+    build_explain_sql, check_token_budget, extract_last_sql_block, extract_table_names,
+    is_write_query, parse_ai_response_segments, strip_sql_fences, AiResponseSegment, AskChoice,
+};
 
-/// Check whether the session token budget has been exceeded.
-///
-/// Returns `true` (and prints a message) if the budget is exceeded,
-/// meaning the caller should abort the AI operation.
-/// Returns `false` if the budget is unlimited (0) or not yet reached.
-fn check_token_budget(settings: &ReplSettings) -> bool {
-    let budget = settings.config.ai.token_budget;
-    if budget == 0 {
-        return false; // No budget limit.
-    }
-    if settings.tokens_used >= budget {
-        eprintln!(
-            "Token budget exhausted ({used}/{budget} tokens used). \
-             AI commands are disabled for this session.",
-            used = settings.tokens_used,
-        );
-        true
-    } else {
-        false
-    }
-}
-
-/// Record token usage from a completion result.
-fn record_token_usage(settings: &mut ReplSettings, result: &crate::ai::CompletionResult) {
-    settings.tokens_used += u64::from(result.input_tokens) + u64::from(result.output_tokens);
-}
-
-/// Prompt the user with a yes/no question and return their answer.
-///
-/// `default_yes` controls what happens when the user presses Enter without
-/// typing anything: `true` → default is yes, `false` → default is no.
-fn ask_yn_prompt(prompt: &str, default_yes: bool) -> bool {
-    use std::io::Write;
-    eprint!("{prompt}");
-    let _ = io::stderr().flush();
-
-    let mut input = String::new();
-    if io::stdin().read_line(&mut input).is_err() {
-        return false;
-    }
-    let answer = input.trim().to_lowercase();
-    if answer.is_empty() {
-        return default_yes;
-    }
-    answer.starts_with('y')
-}
-
-/// User's choice when asked about executing AI-generated SQL.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum AskChoice {
-    /// Execute as-is.
-    Yes,
-    /// Skip execution.
-    No,
-    /// Open in `$EDITOR` first, then execute the edited version.
-    Edit,
-}
-
-/// Prompt the user with `[Y/n/e]` (yes / no / edit) and return their choice.
-///
-/// `default_yes` controls the behaviour when the user presses Enter without
-/// typing: `true` → defaults to `Yes`, `false` → defaults to `No`.
-fn ask_yne_prompt(prompt: &str, default_yes: bool) -> AskChoice {
-    use std::io::Write;
-    eprint!("{prompt}");
-    let _ = io::stderr().flush();
-
-    let mut input = String::new();
-    if io::stdin().read_line(&mut input).is_err() {
-        return AskChoice::No;
-    }
-    let answer = input.trim().to_lowercase();
-    if answer.is_empty() {
-        return if default_yes {
-            AskChoice::Yes
-        } else {
-            AskChoice::No
-        };
-    }
-    if answer.starts_with('e') {
-        AskChoice::Edit
-    } else if answer.starts_with('y') {
-        AskChoice::Yes
-    } else {
-        AskChoice::No
-    }
-}
-
-/// Handle a `/ask <prompt>` command end-to-end.
-///
-/// Acts as a general-purpose `PostgreSQL` expert assistant.  The AI answers
-/// questions directly and, when a database query is needed, includes it in
-/// a triple-backtick `sql` code fence.  Any SQL blocks found in the response
-/// are automatically executed; results are shown interleaved with the AI's
-/// explanatory text.
-///
-/// SQL visibility is controlled by `ai.show_sql = true` (config) or
-/// `\set ECHO_HIDDEN on` (runtime): when either is set the generated SQL is
-/// printed (with syntax highlighting) before its result set.
-///
-/// Write queries (`INSERT`/`UPDATE`/`DELETE`/`MERGE`) always prompt for
-/// confirmation unless running in YOLO mode.
-#[allow(clippy::too_many_lines)]
-async fn handle_ai_ask(
-    client: &Client,
-    prompt: &str,
-    settings: &mut ReplSettings,
-    params: &ConnParams,
-    tx: &mut TxState,
-) {
-    if settings
-        .config
-        .ai
-        .provider
-        .as_deref()
-        .unwrap_or("")
-        .is_empty()
-    {
-        eprintln!(
-            "AI not configured. Add an [ai] section to {}",
-            crate::config::user_config_path_display()
-        );
-        eprintln!("Supported providers: anthropic, openai, ollama");
-        eprintln!("Example:");
-        eprintln!("  [ai]");
-        eprintln!("  provider = \"anthropic\"");
-        eprintln!("  api_key_env = \"ANTHROPIC_API_KEY\"");
-        return;
-    }
-
-    let provider = match get_ai_provider(settings) {
-        Ok(p) => p,
-        Err(e) => {
-            eprintln!("AI error: {e}");
-            return;
-        }
-    };
-
-    // Build a compact schema description for the system prompt.
-    let schema_ctx = match crate::ai::context::build_schema_context(client).await {
-        Ok(ctx) => ctx,
-        Err(e) => {
-            eprintln!("Schema context error: {e}");
-            return;
-        }
-    };
-
-    // Collect wait event context for richer analysis.
-    let wait_ctx = crate::ai::context::build_wait_context(
-        client,
-        settings.db_capabilities.pg_ash.is_available(),
-    )
-    .await;
-
-    let wait_section = if wait_ctx.is_empty() {
-        String::new()
-    } else {
-        format!("\n\nDatabase activity:\n{wait_ctx}")
-    };
-
-    let system_content = format!(
-        "You are a helpful PostgreSQL expert assistant connected to a \
-         live database.\n\
-         Database: {dbname}\n\n\
-         Schema:\n{schema}{wait}\n\n\
-         Guidelines:\n\
-         - Answer the user's question directly and concisely.\n\
-         - If you need to query the database to answer, include the SQL \
-           wrapped in a ```sql code fence.\n\
-         - You may include explanatory text before or after the SQL.\n\
-         - If the question does not require a database query, just answer \
-           it directly — do not generate SQL.\n\
-         - Use standard PostgreSQL syntax.\n\
-         - If the request is ambiguous, make reasonable assumptions.",
-        dbname = params.dbname,
-        schema = schema_ctx,
-        wait = wait_section,
-    );
-
-    // In text2sql mode, strengthen the system prompt so the LLM outputs
-    // only SQL and no surrounding commentary.
-    let system_content = if settings.input_mode == InputMode::Text2Sql {
-        format!(
-            "{system_content}\n\n\
-             IMPORTANT: You are in text2sql mode. Output ONLY the SQL query \
-             inside a ```sql code fence. Do NOT include any explanatory text, \
-             commentary, or description. Just the SQL."
-        )
-    } else {
-        system_content
-    };
-
-    // Build messages: system + conversation history + current prompt.
-    let mut messages = vec![crate::ai::Message {
-        role: crate::ai::Role::System,
-        content: system_content,
-    }];
-
-    // Include conversation history for follow-up context.
-    messages.extend(settings.conversation.to_messages());
-
-    messages.push(crate::ai::Message {
-        role: crate::ai::Role::User,
-        content: prompt.to_owned(),
-    });
-
-    let options = crate::ai::CompletionOptions {
-        model: settings.config.ai.model.clone().unwrap_or_default(),
-        max_tokens: settings.config.ai.max_tokens,
-        temperature: 0.0,
-    };
-
-    let ai_response = match provider.complete(&messages, &options).await {
-        Ok(result) => {
-            record_token_usage(settings, &result);
-            result.content
-        }
-        Err(e) => {
-            eprintln!("AI error: {e}");
-            return;
-        }
-    };
-
-    // Record the exchange in conversation context for follow-ups.
-    settings.conversation.push_user(prompt.to_owned());
-    settings.conversation.push_assistant(ai_response.clone());
-
-    // Auto-compact when approaching the context window limit.
-    if settings
-        .conversation
-        .auto_compact_if_needed(settings.config.ai.context_window)
-    {
-        eprintln!("-- AI context auto-compacted to save tokens");
-    }
-
-    // Parse the response into text and SQL segments, then process each.
-    let show_sql = settings.config.ai.show_sql || settings.echo_hidden;
-    let yolo = settings.exec_mode == ExecMode::Yolo;
-
-    let segments = parse_ai_response_segments(&ai_response);
-
-    for segment in &segments {
-        match segment {
-            AiResponseSegment::Text(text) => {
-                // In text2sql mode, suppress LLM commentary — only show SQL
-                // results. The system prompt already discourages the LLM from
-                // emitting text, but this guard handles any residual output.
-                if settings.input_mode == InputMode::Text2Sql {
-                    continue;
-                }
-                let text = text.trim();
-                if !text.is_empty() {
-                    println!("{text}");
-                }
-            }
-            AiResponseSegment::Sql(sql) => {
-                // Optionally reveal the generated SQL.
-                if show_sql {
-                    if settings.no_highlight {
-                        eprintln!("{sql}");
-                    } else {
-                        eprintln!("{}", crate::highlight::highlight_sql(sql, None));
-                    }
-                }
-
-                let read_only = !is_write_query(sql);
-
-                // In YOLO mode, check autonomy level for write queries.
-                //
-                // L1 Observe     — block all writes (read-only fence).
-                // L2 Supervised  — warn then execute.
-                // L3 Auto        — execute silently.
-                //
-                // --i-know-what-im-doing bypasses all level checks.
-                let yolo_write_action = if yolo && !read_only {
-                    if settings.i_know_what_im_doing {
-                        YoloWriteAction::Execute
-                    } else {
-                        let autonomy = settings
-                            .config
-                            .governance
-                            .autonomy_for(crate::governance::FeatureArea::Rca);
-                        match autonomy {
-                            crate::governance::AutonomyLevel::Observe => YoloWriteAction::Block,
-                            crate::governance::AutonomyLevel::Supervised => {
-                                YoloWriteAction::WarnThenExecute
-                            }
-                            crate::governance::AutonomyLevel::Auto => YoloWriteAction::Execute,
-                        }
-                    }
-                } else {
-                    YoloWriteAction::Execute
-                };
-
-                if yolo_write_action == YoloWriteAction::Block {
-                    eprintln!(
-                        "-- YOLO: write query blocked \
-                         (autonomy level is Observe). \
-                         Use \\set governance.default_autonomy supervised \
-                         or pass --i-know-what-im-doing to bypass"
-                    );
-                    continue;
-                }
-
-                // Read-only queries auto-execute; write queries prompt for
-                // confirmation unless YOLO mode overrides.
-                let auto_exec = read_only || (yolo && yolo_write_action != YoloWriteAction::Block);
-
-                let choice = if auto_exec {
-                    if yolo && !read_only {
-                        match yolo_write_action {
-                            YoloWriteAction::WarnThenExecute => {
-                                eprintln!(
-                                    "-- YOLO: write query executing \
-                                     (autonomy level is Supervised \
-                                     — proceed with care)"
-                                );
-                            }
-                            YoloWriteAction::Execute => {
-                                eprintln!("-- YOLO: auto-executing write query");
-                            }
-                            YoloWriteAction::Block => {
-                                unreachable!("Block handled above")
-                            }
-                        }
-                    }
-                    AskChoice::Yes
-                } else {
-                    ask_yne_prompt("Execute (write query)? [y/N/e] ", false)
-                };
-
-                match choice {
-                    AskChoice::Yes => {
-                        let ok = execute_query(client, sql, settings, tx).await;
-                        if ok {
-                            settings.conversation.push_query_result(sql, "(executed)");
-                        }
-                    }
-                    AskChoice::Edit => match crate::io::edit(sql, None, None) {
-                        Ok(edited) => {
-                            let edited = edited.trim();
-                            if edited.is_empty() {
-                                eprintln!("(empty — skipped)");
-                            } else {
-                                let ok = execute_query(client, edited, settings, tx).await;
-                                if ok {
-                                    settings
-                                        .conversation
-                                        .push_query_result(edited, "(executed after edit)");
-                                }
-                            }
-                        }
-                        Err(e) => eprintln!("{e}"),
-                    },
-                    AskChoice::No => {}
-                }
-            }
-        }
-    }
-}
-
-/// A segment of an AI response: plain text or a SQL block.
-enum AiResponseSegment {
-    Text(String),
-    Sql(String),
-}
-
-/// Split an AI response into alternating text and SQL segments.
-///
-/// SQL blocks are delimited by ` ```sql ` … ` ``` ` (case-insensitive
-/// language tag).  Plain ` ``` ` fences without a language tag are treated
-/// as text, not SQL.  The function never allocates an empty segment.
-fn parse_ai_response_segments(response: &str) -> Vec<AiResponseSegment> {
-    let mut segments: Vec<AiResponseSegment> = Vec::new();
-    let mut remaining = response;
-
-    while !remaining.is_empty() {
-        // Look for the start of a ```sql fence.
-        if let Some(fence_start) = remaining.find("```sql") {
-            // Any text before the fence becomes a Text segment.
-            let before = &remaining[..fence_start];
-            if !before.trim().is_empty() {
-                segments.push(AiResponseSegment::Text(before.to_owned()));
-            }
-            // Advance past the opening fence + language tag.
-            let after_open = &remaining[fence_start + 6..];
-            // Skip the newline immediately after the tag.
-            let body_start = after_open
-                .find('\n')
-                .map_or(after_open, |i| &after_open[i + 1..]);
-            // Find the closing fence.
-            if let Some(close_pos) = body_start.find("```") {
-                let sql_body = body_start[..close_pos].trim();
-                if !sql_body.is_empty() {
-                    segments.push(AiResponseSegment::Sql(sql_body.to_owned()));
-                }
-                remaining = &body_start[close_pos + 3..];
-            } else {
-                // Unclosed fence: treat everything as SQL.
-                let sql_body = body_start.trim();
-                if !sql_body.is_empty() {
-                    segments.push(AiResponseSegment::Sql(sql_body.to_owned()));
-                }
-                break;
-            }
-        } else {
-            // No more SQL fences — rest is plain text.
-            if !remaining.trim().is_empty() {
-                segments.push(AiResponseSegment::Text(remaining.to_owned()));
-            }
-            break;
-        }
-    }
-
-    segments
-}
-
-/// Handle a plan-mode prompt.
-///
-/// Gathers schema context, sends the user's natural-language prompt to the
-/// LLM with a plan-generation system prompt, and streams the resulting plan.
-/// Offers to save the plan to `~/.local/share/rpg/plans/`.
-#[allow(clippy::too_many_lines)]
-async fn handle_ai_plan(
-    client: &Client,
-    prompt: &str,
-    settings: &ReplSettings,
-    params: &ConnParams,
-) {
-    if settings
-        .config
-        .ai
-        .provider
-        .as_deref()
-        .unwrap_or("")
-        .is_empty()
-    {
-        eprintln!(
-            "AI not configured. Add an [ai] section to {}",
-            crate::config::user_config_path_display()
-        );
-        return;
-    }
-
-    let provider = match get_ai_provider(settings) {
-        Ok(p) => p,
-        Err(e) => {
-            eprintln!("AI error: {e}");
-            return;
-        }
-    };
-
-    let schema_ctx = match crate::ai::context::build_schema_context(client).await {
-        Ok(ctx) => ctx,
-        Err(e) => {
-            eprintln!("Schema context error: {e}");
-            return;
-        }
-    };
-
-    // Collect live diagnostic data (read-only queries) for richer context.
-    eprintln!("-- Plan mode: collecting diagnostics...");
-    let pg_ash = settings.db_capabilities.pg_ash.is_available();
-    let diagnostics = crate::rca::collect_snapshot(client, pg_ash).await;
-    let diag_steps = diagnostics.steps.iter().filter(|s| s.has_data).count();
-    let diag_context = diagnostics.to_prompt();
-    eprintln!("-- Collected {diag_steps} diagnostic steps");
-
-    let system_content = format!(
-        "You are a PostgreSQL expert. \
-         The user has asked you to investigate and produce an action plan.\n\
-         Database: {dbname}\n\n\
-         Schema:\n{schema}\n\n\
-         Live diagnostics (read-only queries just collected):\n{diagnostics}\n\n\
-         Rules:\n\
-         - Produce a structured plan in markdown format\n\
-         - Each action should include the SQL command and a safety assessment\n\
-         - Mark actions as [safe], [caution], or [dangerous]\n\
-         - Order actions from safest to most impactful\n\
-         - Include estimated duration where possible\n\
-         - Start with a brief root-cause analysis based on the live diagnostics above\n\
-         - Reference specific data from the diagnostics to support your analysis\n\
-         - Do NOT execute anything — only plan",
-        dbname = params.dbname,
-        schema = schema_ctx,
-        diagnostics = diag_context,
-    );
-
-    let messages = vec![
-        crate::ai::Message {
-            role: crate::ai::Role::System,
-            content: system_content,
-        },
-        crate::ai::Message {
-            role: crate::ai::Role::User,
-            content: prompt.to_owned(),
-        },
-    ];
-
-    let options = crate::ai::CompletionOptions {
-        model: settings.config.ai.model.clone().unwrap_or_default(),
-        max_tokens: settings.config.ai.max_tokens,
-        temperature: 0.0,
-    };
-
-    eprintln!("-- Plan mode: investigating...");
-    let result = match stream_completion(
-        provider.as_ref(),
-        &messages,
-        &options,
-        settings.no_highlight,
-    )
-    .await
-    {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("AI error: {e}");
-            return;
-        }
-    };
-
-    // Offer to save the plan.
-    if ask_yn_prompt("Save this plan? [y/N] ", false) {
-        let plans_dir = dirs::data_local_dir()
-            .unwrap_or_else(|| std::path::PathBuf::from("."))
-            .join("rpg")
-            .join("plans");
-        if let Err(e) = std::fs::create_dir_all(&plans_dir) {
-            eprintln!("Cannot create plans directory: {e}");
-            return;
-        }
-        let date = format_system_time(std::time::SystemTime::now())
-            .replace(' ', "-")
-            .replace(':', "");
-        // Build a slug from the first few words of the prompt.
-        let slug: String = prompt
-            .split_whitespace()
-            .take(4)
-            .collect::<Vec<_>>()
-            .join("-")
-            .chars()
-            .filter(|c| c.is_alphanumeric() || *c == '-')
-            .collect();
-        let filename = format!("{date}-{slug}.md");
-        let path = plans_dir.join(&filename);
-        match std::fs::write(&path, &result.content) {
-            Ok(()) => eprintln!("Saved to: {}", path.display()),
-            Err(e) => eprintln!("Failed to save plan: {e}"),
-        }
-    }
-}
-
-/// Extract the last SQL code block from a mixed text+SQL LLM response.
-///
-/// LLMs responding to `/fix` produce a mix of explanation and corrected SQL.
-/// This function scans for the last `` ```sql ... ``` `` fence (or plain
-/// `` ``` ... ``` `` fence) and returns the inner content, trimmed.  If no
-/// fences are found it returns `None`.
-fn extract_last_sql_block(text: &str) -> Option<&str> {
-    // Iterate forward through all fence pairs, keeping the last non-empty one.
-    // rfind("```") was incorrect: it found the closing fence, then looked for
-    // another "```" after it — always returning None for a single block.
-    let mut last_sql: Option<&str> = None;
-    let mut search_from = 0;
-    while let Some(open_pos) = text[search_from..].find("```") {
-        let open_abs = search_from + open_pos;
-        let after_open = &text[open_abs + 3..];
-        // Skip optional language tag (e.g. "sql\n") on the opening fence line.
-        let body_start_rel = after_open.find('\n').map_or(after_open.len(), |i| i + 1);
-        let body_text = &after_open[body_start_rel..];
-        if let Some(close_pos) = body_text.find("```") {
-            let body = body_text[..close_pos].trim();
-            if !body.is_empty() {
-                last_sql = Some(body);
-            }
-            // Advance past the closing fence before looking for the next pair.
-            search_from = open_abs + 3 + body_start_rel + close_pos + 3;
-        } else {
-            // Unclosed fence — treat rest of text as the body.
-            let body = body_text.trim();
-            if !body.is_empty() {
-                last_sql = Some(body);
-            }
-            break;
-        }
-    }
-    last_sql
-}
-
-/// Handle a `/fix` command end-to-end.
-///
-/// Looks up the most recently failed query from [`ReplSettings::last_error`],
-/// sends it to the configured LLM with schema context, and prints an
-/// explanation plus a corrected SQL query.  After streaming the response,
-/// if a corrected SQL block is detected the user is prompted
-/// `Execute? [Y/n]` (default yes) and the query is executed when confirmed.
-/// Gracefully degrades when no prior error exists or when AI is not
-/// configured.
-#[allow(clippy::too_many_lines)]
-async fn handle_ai_fix(
-    client: &Client,
-    settings: &mut ReplSettings,
-    params: &ConnParams,
-    tx: &mut TxState,
-) {
-    // Require a prior error to fix.
-    let last_error = if let Some(e) = &settings.last_error {
-        e.clone()
-    } else {
-        eprintln!("No recent error to fix. Run a query first.");
-        return;
-    };
-
-    if settings
-        .config
-        .ai
-        .provider
-        .as_deref()
-        .unwrap_or("")
-        .is_empty()
-    {
-        eprintln!(
-            "AI not configured. Add an [ai] section to {}",
-            crate::config::user_config_path_display()
-        );
-        eprintln!("Supported providers: anthropic, openai, ollama");
-        eprintln!("Example:");
-        eprintln!("  [ai]");
-        eprintln!("  provider = \"anthropic\"");
-        eprintln!("  api_key_env = \"ANTHROPIC_API_KEY\"");
-        return;
-    }
-
-    let provider = match get_ai_provider(settings) {
-        Ok(p) => p,
-        Err(e) => {
-            eprintln!("AI error: {e}");
-            return;
-        }
-    };
-
-    // Build a compact schema description for the system prompt.
-    let schema_ctx = match crate::ai::context::build_schema_context(client).await {
-        Ok(ctx) => ctx,
-        Err(e) => {
-            eprintln!("Schema context error: {e}");
-            return;
-        }
-    };
-
-    // Format the SQLSTATE hint if available.
-    let sqlstate_hint = last_error
-        .sqlstate
-        .as_deref()
-        .map(|s| format!(" (SQLSTATE {s})"))
-        .unwrap_or_default();
-
-    let system_content = format!(
-        "You are a PostgreSQL expert. \
-         Explain SQL errors and provide corrected queries.\n\
-         Database: {dbname}\n\n\
-         Schema:\n{schema}\n\n\
-         Rules:\n\
-         - First, briefly explain what caused the error (1-2 sentences)\n\
-         - Then output the corrected SQL query\n\
-         - Use standard PostgreSQL syntax\n\
-         - Keep the corrected query as close to the original intent as possible",
-        dbname = params.dbname,
-        schema = schema_ctx,
-    );
-
-    let user_content = format!(
-        "The following query failed{sqlstate_hint}:\n\n\
-         ```sql\n{query}\n```\n\n\
-         Error: {error}",
-        query = last_error.query,
-        error = last_error.error_message,
-    );
-
-    let messages = vec![
-        crate::ai::Message {
-            role: crate::ai::Role::System,
-            content: system_content,
-        },
-        crate::ai::Message {
-            role: crate::ai::Role::User,
-            content: user_content,
-        },
-    ];
-
-    let options = crate::ai::CompletionOptions {
-        model: settings.config.ai.model.clone().unwrap_or_default(),
-        max_tokens: settings.config.ai.max_tokens,
-        temperature: 0.0,
-    };
-
-    let result = match stream_completion(
-        provider.as_ref(),
-        &messages,
-        &options,
-        settings.no_highlight,
-    )
-    .await
-    {
-        Ok(r) => r,
-        Err(e) => {
-            eprintln!("AI error: {e}");
-            return;
-        }
-    };
-    record_token_usage(settings, &result);
-
-    // If the response contains a corrected SQL block, offer to execute it.
-    if let Some(fix_sql) = extract_last_sql_block(&result.content) {
-        let choice = ask_yne_prompt("Execute? [Y/n/e] ", true);
-        match choice {
-            AskChoice::Yes => {
-                execute_query(client, fix_sql, settings, tx).await;
-            }
-            AskChoice::Edit => match crate::io::edit(fix_sql, None, None) {
-                Ok(edited) => {
-                    let edited = edited.trim();
-                    if edited.is_empty() {
-                        eprintln!("(empty — skipped)");
-                    } else {
-                        execute_query(client, edited, settings, tx).await;
-                    }
-                }
-                Err(e) => eprintln!("{e}"),
-            },
-            AskChoice::No => {}
-        }
-    }
-}
-
-/// Detect whether a query is a data-modifying statement that must be
-/// wrapped in a rolled-back transaction before `EXPLAIN ANALYZE`.
-///
-/// Returns `true` for `INSERT`, `UPDATE`, `DELETE`, and `MERGE`.
-fn is_write_query(sql: &str) -> bool {
-    let first = sql.split_whitespace().next().unwrap_or("").to_uppercase();
-    matches!(first.as_str(), "INSERT" | "UPDATE" | "DELETE" | "MERGE")
-}
-
-/// Build the `EXPLAIN` SQL for a given target query.
-///
-/// Write queries are wrapped in `BEGIN` / `ROLLBACK` so that
-/// `EXPLAIN ANALYZE` can run them without persisting any changes.
-fn build_explain_sql(target_query: &str) -> String {
-    let explain = format!("explain (analyze, costs, verbose, buffers, format text) {target_query}");
-    if is_write_query(target_query) {
-        format!("begin;\n{explain};\nrollback;")
-    } else {
-        explain
-    }
-}
-
-/// Handle a `/explain [query]` command end-to-end.
-///
-/// 1. Resolves the target query: inline arg or `last_query`.
-/// 2. Runs `EXPLAIN (ANALYZE, COSTS, VERBOSE, BUFFERS, FORMAT TEXT)`.
-///    Write queries (`INSERT`/`UPDATE`/`DELETE`/`MERGE`) are wrapped in
-///    a `BEGIN` … `ROLLBACK` to prevent side-effects.
-/// 3. Prints the raw plan.
-/// 4. If AI is configured, sends plan + schema context to the LLM and
-///    prints its interpretation.
-#[allow(clippy::too_many_lines)]
-async fn handle_ai_explain(
-    client: &Client,
-    query_arg: &str,
-    settings: &mut ReplSettings,
-    params: &ConnParams,
-) {
-    // Resolve target query.
-    let target_query = if query_arg.is_empty() {
-        if let Some(q) = settings.last_query.as_deref() {
-            q.to_owned()
-        } else {
-            eprintln!(
-                "/explain: no query to explain. \
-                 Run a query first or provide one: /explain SELECT ..."
-            );
-            return;
-        }
-    } else {
-        query_arg.to_owned()
-    };
-
-    // Run EXPLAIN ANALYZE (wrapped in BEGIN/ROLLBACK for write queries).
-    let explain_sql = build_explain_sql(&target_query);
-
-    let messages_result = client.simple_query(&explain_sql).await;
-    let raw_messages = match messages_result {
-        Ok(msgs) => msgs,
-        Err(e) => {
-            crate::output::eprint_db_error(&e, Some(&target_query), settings.verbose_errors);
-            return;
-        }
-    };
-
-    // Collect plan lines from the result.
-    let mut plan_lines: Vec<String> = Vec::new();
-    for msg in &raw_messages {
-        if let tokio_postgres::SimpleQueryMessage::Row(row) = msg {
-            if let Some(line) = row.get(0) {
-                plan_lines.push(line.to_owned());
-            }
-        }
-    }
-
-    if plan_lines.is_empty() {
-        eprintln!("/explain: EXPLAIN returned no output");
-        return;
-    }
-
-    let plan_text = plan_lines.join("\n");
-    println!("{plan_text}");
-
-    // AI interpretation — skip gracefully when AI is not configured.
-    let Ok(provider) = get_ai_provider(settings) else {
-        return;
-    };
-
-    // Build schema context for richer analysis.
-    let schema_ctx = match crate::ai::context::build_schema_context(client).await {
-        Ok(ctx) => ctx,
-        Err(e) => {
-            eprintln!("Schema context error: {e}");
-            return;
-        }
-    };
-
-    // Collect wait event context for performance correlation.
-    let wait_ctx = crate::ai::context::build_wait_context(
-        client,
-        settings.db_capabilities.pg_ash.is_available(),
-    )
-    .await;
-
-    let wait_section = if wait_ctx.is_empty() {
-        String::new()
-    } else {
-        format!(
-            "\n\nDatabase activity (use to correlate plan behavior \
-             with current wait patterns):\n{wait_ctx}"
-        )
-    };
-
-    let system_content = format!(
-        "You are a PostgreSQL performance expert. \
-         Analyse the EXPLAIN ANALYZE plan provided by the user and give \
-         a concise, actionable interpretation:\n\
-         - Identify the most expensive nodes\n\
-         - Flag sequential scans on large tables\n\
-         - Note any high row-estimate errors\n\
-         - Suggest specific indexes or query rewrites when applicable\n\
-         - If wait event data is provided, correlate plan behavior with waits\n\n\
-         Database: {dbname}\n\n\
-         Schema:\n{schema}{wait}",
-        dbname = params.dbname,
-        schema = schema_ctx,
-        wait = wait_section,
-    );
-
-    let user_content = format!("Query:\n{target_query}\n\nEXPLAIN ANALYZE output:\n{plan_text}");
-
-    let ai_messages = vec![
-        crate::ai::Message {
-            role: crate::ai::Role::System,
-            content: system_content,
-        },
-        crate::ai::Message {
-            role: crate::ai::Role::User,
-            content: user_content,
-        },
-    ];
-
-    let options = crate::ai::CompletionOptions {
-        model: settings.config.ai.model.clone().unwrap_or_default(),
-        max_tokens: settings.config.ai.max_tokens,
-        temperature: 0.0,
-    };
-
-    println!();
-    match stream_completion(
-        provider.as_ref(),
-        &ai_messages,
-        &options,
-        settings.no_highlight,
-    )
-    .await
-    {
-        Ok(result) => record_token_usage(settings, &result),
-        Err(e) => eprintln!("AI error: {e}"),
-    }
-}
-
-/// Extract table names referenced by `FROM` and `JOIN` clauses.
-///
-/// Best-effort heuristic parser — handles common patterns including
-/// schema-qualified names but does not aim for full SQL parsing.
-/// Used by `/optimize` to query `pg_stat_user_tables`.
-fn extract_table_names(sql: &str) -> Vec<String> {
-    let upper = sql.to_uppercase();
-    let tokens: Vec<&str> = sql.split_whitespace().collect();
-    let upper_tokens: Vec<String> = upper.split_whitespace().map(String::from).collect();
-    let mut tables = Vec::new();
-
-    let mut i = 0;
-    while i < upper_tokens.len() {
-        let is_from = upper_tokens[i] == "FROM";
-        let is_join = upper_tokens[i].ends_with("JOIN") && upper_tokens[i] != "DISJOIN";
-
-        if (is_from || is_join) && i + 1 < tokens.len() {
-            let candidate = tokens[i + 1];
-            // Skip sub-selects: FROM (SELECT ...)
-            if !candidate.starts_with('(') {
-                let clean = candidate.trim_end_matches([',', ')', ';']);
-                if !clean.is_empty() {
-                    tables.push(clean.to_owned());
-                }
-            }
-        }
-        i += 1;
-    }
-
-    tables.sort();
-    tables.dedup();
-    tables
-}
-
-/// Handle a `/optimize [query]` command end-to-end.
-///
-/// 1. Resolves the target query: inline arg or `last_query`.
-/// 2. Runs `EXPLAIN (ANALYZE, COSTS, VERBOSE, BUFFERS, FORMAT TEXT)`.
-/// 3. Gathers `pg_stat_user_tables` stats for referenced tables.
-/// 4. Sends plan + stats + schema context to the LLM for optimization
-///    suggestions (index creation, query rewrites, join order changes).
-#[allow(clippy::too_many_lines)]
-async fn handle_ai_optimize(
-    client: &Client,
-    query_arg: &str,
-    settings: &mut ReplSettings,
-    params: &ConnParams,
-) {
-    // Resolve target query.
-    let target_query = if query_arg.is_empty() {
-        if let Some(q) = settings.last_query.as_deref() {
-            q.to_owned()
-        } else {
-            eprintln!(
-                "/optimize: no query to optimize. \
-                 Run a query first or provide one: /optimize SELECT ..."
-            );
-            return;
-        }
-    } else {
-        query_arg.to_owned()
-    };
-
-    // Run EXPLAIN ANALYZE (wrapped in BEGIN/ROLLBACK for write queries).
-    let explain_sql = build_explain_sql(&target_query);
-
-    let raw_messages = match client.simple_query(&explain_sql).await {
-        Ok(msgs) => msgs,
-        Err(e) => {
-            crate::output::eprint_db_error(&e, Some(&target_query), settings.verbose_errors);
-            return;
-        }
-    };
-
-    // Collect plan lines.
-    let mut plan_lines: Vec<String> = Vec::new();
-    for msg in &raw_messages {
-        if let tokio_postgres::SimpleQueryMessage::Row(row) = msg {
-            if let Some(line) = row.get(0) {
-                plan_lines.push(line.to_owned());
-            }
-        }
-    }
-
-    if plan_lines.is_empty() {
-        eprintln!("/optimize: EXPLAIN returned no output");
-        return;
-    }
-
-    let plan_text = plan_lines.join("\n");
-    println!("{plan_text}");
-
-    // Gather table statistics for referenced tables.
-    let table_names = extract_table_names(&target_query);
-    let mut stats_text = String::new();
-
-    if !table_names.is_empty() {
-        let in_list: String = table_names
-            .iter()
-            .map(|t| {
-                let escaped = t.replace('\'', "''");
-                format!("'{escaped}'")
-            })
-            .collect::<Vec<_>>()
-            .join(", ");
-
-        let stats_sql = format!(
-            "SELECT schemaname || '.' || relname AS table_name, \
-                    n_live_tup, n_dead_tup, \
-                    seq_scan, seq_tup_read, \
-                    idx_scan, idx_tup_fetch, \
-                    last_vacuum::text, last_analyze::text \
-             FROM pg_stat_user_tables \
-             WHERE relname IN ({in_list}) \
-                OR schemaname || '.' || relname IN ({in_list}) \
-             ORDER BY relname"
-        );
-
-        if let Ok(msgs) = client.simple_query(&stats_sql).await {
-            let mut stat_rows = Vec::new();
-            for msg in &msgs {
-                if let tokio_postgres::SimpleQueryMessage::Row(row) = msg {
-                    let cols: Vec<String> = (0..9)
-                        .map(|i| row.get(i).unwrap_or("(null)").to_owned())
-                        .collect();
-                    stat_rows.push(cols.join(" | "));
-                }
-            }
-            if !stat_rows.is_empty() {
-                stats_text = format!(
-                    "\n\nTable statistics (table | live_tup | dead_tup | \
-                     seq_scan | seq_tup_read | idx_scan | idx_tup_fetch | \
-                     last_vacuum | last_analyze):\n{}",
-                    stat_rows.join("\n")
-                );
-            }
-        }
-    }
-
-    // AI optimization — skip gracefully when AI is not configured.
-    let Ok(provider) = get_ai_provider(settings) else {
-        eprintln!(
-            "\nAI not configured — showing raw plan only. \
-             Add an [ai] section to {} for optimization suggestions.",
-            crate::config::user_config_path_display()
-        );
-        return;
-    };
-
-    let schema_ctx = match crate::ai::context::build_schema_context(client).await {
-        Ok(ctx) => ctx,
-        Err(e) => {
-            eprintln!("Schema context error: {e}");
-            return;
-        }
-    };
-
-    let system_content = format!(
-        "You are a PostgreSQL performance optimization expert. \
-         Analyse the query, its EXPLAIN ANALYZE plan, and table statistics, \
-         then provide actionable optimization suggestions.\n\
-         Database: {dbname}\n\n\
-         Schema:\n{schema}\n\n\
-         Rules:\n\
-         - Identify the most expensive operations in the plan\n\
-         - Suggest specific CREATE INDEX statements when beneficial\n\
-         - Suggest query rewrites (join order, CTEs, subquery elimination)\n\
-         - Note any sequential scans on large tables\n\
-         - Estimate the expected improvement for each suggestion\n\
-         - Output suggestions ordered by expected impact (highest first)",
-        dbname = params.dbname,
-        schema = schema_ctx,
-    );
-
-    let user_content = format!(
-        "Query:\n```sql\n{target_query}\n```\n\n\
-         EXPLAIN ANALYZE output:\n{plan_text}{stats_text}"
-    );
-
-    let ai_messages = vec![
-        crate::ai::Message {
-            role: crate::ai::Role::System,
-            content: system_content,
-        },
-        crate::ai::Message {
-            role: crate::ai::Role::User,
-            content: user_content,
-        },
-    ];
-
-    let options = crate::ai::CompletionOptions {
-        model: settings.config.ai.model.clone().unwrap_or_default(),
-        max_tokens: settings.config.ai.max_tokens,
-        temperature: 0.0,
-    };
-
-    println!();
-    match stream_completion(
-        provider.as_ref(),
-        &ai_messages,
-        &options,
-        settings.no_highlight,
-    )
-    .await
-    {
-        Ok(result) => record_token_usage(settings, &result),
-        Err(e) => eprintln!("AI error: {e}"),
-    }
-}
-
-/// Handle a `/describe <table>` command.
-///
-/// Queries the table's columns, constraints, indexes, and row estimate,
-/// then sends everything to the LLM for a human-readable description of
-/// the table's purpose, relationships, and notable patterns.
-#[allow(clippy::too_many_lines)]
-async fn handle_ai_describe(
-    client: &Client,
-    table_name: &str,
-    settings: &mut ReplSettings,
-    params: &ConnParams,
-) {
-    let Ok(provider) = get_ai_provider(settings) else {
-        eprintln!(
-            "AI not configured. Add an [ai] section to {}",
-            crate::config::user_config_path_display()
-        );
-        return;
-    };
-
-    // Gather table metadata.
-    let mut table_info = String::new();
-
-    // Columns.
-    let col_query = format!(
-        "SELECT column_name, data_type, is_nullable, column_default \
-         FROM information_schema.columns \
-         WHERE table_name = '{table_name}' \
-         ORDER BY ordinal_position"
-    );
-    if let Ok(rows) = client.simple_query(&col_query).await {
-        use std::fmt::Write as _;
-        let _ = writeln!(table_info, "Columns:");
-        for msg in &rows {
-            if let tokio_postgres::SimpleQueryMessage::Row(row) = msg {
-                let name = row.get(0).unwrap_or("?");
-                let dtype = row.get(1).unwrap_or("?");
-                let nullable = row.get(2).unwrap_or("?");
-                let default = row.get(3).unwrap_or("");
-                let _ = writeln!(
-                    table_info,
-                    "  {name} {dtype} nullable={nullable} default={default}"
-                );
-            }
-        }
-    }
-
-    // Constraints (PK, FK, unique, check).
-    let constraint_query = format!(
-        "SELECT conname, contype, pg_get_constraintdef(oid) \
-         FROM pg_constraint \
-         WHERE conrelid = '{table_name}'::regclass"
-    );
-    if let Ok(rows) = client.simple_query(&constraint_query).await {
-        use std::fmt::Write as _;
-        let _ = writeln!(table_info, "\nConstraints:");
-        for msg in &rows {
-            if let tokio_postgres::SimpleQueryMessage::Row(row) = msg {
-                let name = row.get(0).unwrap_or("?");
-                let ctype = row.get(1).unwrap_or("?");
-                let def = row.get(2).unwrap_or("?");
-                let type_label = match ctype {
-                    "p" => "PRIMARY KEY",
-                    "f" => "FOREIGN KEY",
-                    "u" => "UNIQUE",
-                    "c" => "CHECK",
-                    "x" => "EXCLUSION",
-                    other => other,
-                };
-                let _ = writeln!(table_info, "  {name} ({type_label}): {def}");
-            }
-        }
-    }
-
-    // Indexes.
-    let idx_query = format!(
-        "SELECT indexname, indexdef \
-         FROM pg_indexes \
-         WHERE tablename = '{table_name}'"
-    );
-    if let Ok(rows) = client.simple_query(&idx_query).await {
-        use std::fmt::Write as _;
-        let _ = writeln!(table_info, "\nIndexes:");
-        for msg in &rows {
-            if let tokio_postgres::SimpleQueryMessage::Row(row) = msg {
-                let name = row.get(0).unwrap_or("?");
-                let def = row.get(1).unwrap_or("?");
-                let _ = writeln!(table_info, "  {name}: {def}");
-            }
-        }
-    }
-
-    // Row estimate + size.
-    let stats_query = format!(
-        "SELECT reltuples::bigint AS row_estimate, \
-         pg_size_pretty(pg_total_relation_size('{table_name}'::regclass)) AS size \
-         FROM pg_class WHERE relname = '{table_name}'"
-    );
-    if let Ok(rows) = client.simple_query(&stats_query).await {
-        use std::fmt::Write as _;
-        for msg in &rows {
-            if let tokio_postgres::SimpleQueryMessage::Row(row) = msg {
-                let rows_est = row.get(0).unwrap_or("?");
-                let size = row.get(1).unwrap_or("?");
-                let _ = writeln!(
-                    table_info,
-                    "\nEstimated rows: {rows_est}, Total size: {size}"
-                );
-            }
-        }
-    }
-
-    if table_info.trim().is_empty() {
-        eprintln!("No metadata found for table '{table_name}'.");
-        return;
-    }
-
-    let system_content = format!(
-        "You are a PostgreSQL expert. \
-         Describe the purpose and design of this database table.\n\
-         Database: {dbname}\n\n\
-         Rules:\n\
-         - Infer the table's purpose from its name, columns, and constraints\n\
-         - Describe relationships (foreign keys) to other tables\n\
-         - Note any design patterns (audit columns, soft deletes, etc.)\n\
-         - Mention notable indexes and their likely purpose\n\
-         - Be concise — this is for quick understanding",
-        dbname = params.dbname,
-    );
-
-    let messages = vec![
-        crate::ai::Message {
-            role: crate::ai::Role::System,
-            content: system_content,
-        },
-        crate::ai::Message {
-            role: crate::ai::Role::User,
-            content: format!("Describe table '{table_name}':\n\n{table_info}"),
-        },
-    ];
-
-    let options = crate::ai::CompletionOptions {
-        model: settings.config.ai.model.clone().unwrap_or_default(),
-        max_tokens: settings.config.ai.max_tokens,
-        temperature: 0.0,
-    };
-
-    match stream_completion(
-        provider.as_ref(),
-        &messages,
-        &options,
-        settings.no_highlight,
-    )
-    .await
-    {
-        Ok(result) => record_token_usage(settings, &result),
-        Err(e) => eprintln!("AI error: {e}"),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// /rca — Root Cause Analysis
-// ---------------------------------------------------------------------------
-
-/// Handle the `/rca` command: collect diagnostic snapshot and analyze.
-async fn handle_ai_rca(client: &Client, settings: &mut ReplSettings, params: &ConnParams) {
-    // Check whether AI is configured; if not, still show raw diagnostic data.
-    let ai_configured = settings
-        .config
-        .ai
-        .provider
-        .as_deref()
-        .is_some_and(|p| !p.is_empty());
-
-    if !ai_configured {
-        // Without AI, still collect and display the diagnostic snapshot.
-        eprintln!("AI not configured — collecting raw diagnostic data only.");
-        let pg_ash = settings
-            .config
-            .governance
-            .autonomy_for(crate::governance::FeatureArea::Rca);
-        let _ = pg_ash; // autonomy level not used in Observe mode yet
-        let snapshot = crate::rca::collect_snapshot(client, false).await;
-        print!("{}", snapshot.to_prompt());
-        return;
-    }
-
-    eprintln!("Collecting diagnostic data...");
-
-    // Detect pg_ash availability from capabilities (set at connect time).
-    let pg_ash_available = settings.db_capabilities.pg_ash.is_available();
-
-    let snapshot = crate::rca::collect_snapshot(client, pg_ash_available).await;
-
-    // Show raw data summary.
-    let data_steps = snapshot.steps.iter().filter(|s| s.has_data).count();
-    let total_steps = snapshot.steps.len();
-    eprintln!("Collected {data_steps}/{total_steps} steps with data. Analyzing...\n");
-
-    let provider = match get_ai_provider(settings) {
-        Ok(p) => p,
-        Err(e) => {
-            eprintln!("AI error: {e}");
-            return;
-        }
-    };
-
-    // Build schema context for richer analysis.
-    let schema_ctx = match crate::ai::context::build_schema_context(client).await {
-        Ok(ctx) => ctx,
-        Err(e) => {
-            eprintln!("Schema context error: {e}");
-            String::new()
-        }
-    };
-
-    let system_content = crate::rca::rca_system_prompt(&schema_ctx);
-    let user_content = format!(
-        "Database: {dbname}\n\n{snapshot}",
-        dbname = params.dbname,
-        snapshot = snapshot.to_prompt(),
-    );
-
-    let messages = vec![
-        crate::ai::Message {
-            role: crate::ai::Role::System,
-            content: system_content,
-        },
-        crate::ai::Message {
-            role: crate::ai::Role::User,
-            content: user_content,
-        },
-    ];
-
-    let options = crate::ai::CompletionOptions {
-        model: settings.config.ai.model.clone().unwrap_or_default(),
-        max_tokens: settings.config.ai.max_tokens,
-        temperature: 0.0,
-    };
-
-    match stream_completion(
-        provider.as_ref(),
-        &messages,
-        &options,
-        settings.no_highlight,
-    )
-    .await
-    {
-        Ok(result) => record_token_usage(settings, &result),
-        Err(e) => eprintln!("AI error: {e}"),
-    }
-
-    // In Supervised mode, propose actionable mitigations after analysis.
-    let rca_autonomy = settings
-        .config
-        .governance
-        .autonomy_for(crate::governance::FeatureArea::Rca);
-    if rca_autonomy == crate::governance::AutonomyLevel::Supervised {
-        let proposals = crate::rca_actions::propose_mitigations(client).await;
-        if !proposals.is_empty() {
-            let mut audit_log = crate::governance::AuditLog::new();
-            crate::rca_actions::run_supervised_flow(client, &proposals, &mut audit_log).await;
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
 // Unit tests
 // ---------------------------------------------------------------------------
